Move rsync into its own module.
