+++ src/repository.rs
@@ -6,18 +6,14 @@
 //!
 //! [`Repository`]: struct.Repository.html
 
-use std::{fs, io, mem, process};
-use std::collections::HashMap;
-use std::fs::{DirEntry, File, create_dir_all};
-use std::io::Read;
+use std::{fs, io};
+use std::fs::File;
 use std::path::{Path, PathBuf};
-use std::process::{ExitStatus, Output};
-use std::sync::{Arc, Condvar, Mutex};
-use std::time::{Duration, SystemTime, SystemTimeError};
+use std::sync::Arc;
 use bytes::Bytes;
 use derive_more::Display;
 use futures::future;
-use futures::{Future, IntoFuture};
+use futures::Future;
 use futures_cpupool::CpuPool;
 use log::{debug, error, info, warn};
 use rpki::uri;
@@ -27,7 +23,8 @@ use rpki::manifest::{Manifest, ManifestContent, ManifestHash};
 use rpki::roa::Roa;
 use rpki::tal::Tal;
 use rpki::x509::ValidationError;
-use unwrap::unwrap;
+use crate::rsync;
+use crate::metrics::Metrics;
 use crate::config::Config;
 use crate::operation::Error;
 use crate::origins::RouteOrigins;
@@ -62,8 +59,8 @@ pub struct Repository(Arc<RepoInner>);
 
 #[derive(Debug)]
 struct RepoInner {
-    /// The directory our local copy of the respository lives in.
-    cache_dir: PathBuf, 
+    /// The base directory of the local cache.
+    cache_dir: PathBuf,
 
     /// The list of our TALs. 
     tals: Vec<Tal>,
@@ -80,10 +77,8 @@ struct RepoInner {
     /// Number of validation threads.
     validation_threads: usize,
 
-    /// Information for running rsync.
-    ///
-    /// If this is `None`, we don’t rsync.
-    rsync: Option<(Mutex<RsyncState>, RsyncCommand)>,
+    /// The rsync cache.
+    rsync: rsync::Cache,
 
     /// Should we leave the repository dirty after a valiation run.
     dirty_repository: bool,
@@ -92,12 +87,12 @@ struct RepoInner {
 impl Repository {
     /// Creates a new repository.
     ///
-    /// Takes all necessary information from `config`. If `rsync` is `false`,
-    /// rsyncing is disabled.
+    /// Takes all necessary information from `config`. If `update` is `false`,
+    /// updating the local cache will not be updated from upstream.
     pub fn new(
         config: &Config,
         extra_output: bool,
-        rsync: bool
+        update: bool
     ) -> Result<Self, Error> {
         if let Err(err) = fs::read_dir(&config.cache_dir) {
             if err.kind() == io::ErrorKind::NotFound {
@@ -124,15 +119,10 @@ impl Repository {
             extra_output,
             rsync_threads: config.rsync_count,
             validation_threads: config.validation_threads,
-            rsync: if rsync {
-                Some((
-                    Mutex::new(RsyncState::new()),
-                    RsyncCommand::detect(config)?
-                ))
-            }
-            else {
-                None
-            },
+            rsync: rsync::Cache::new(
+                config, config.cache_dir.join("rsync"), update
+            
+            )?,
             dirty_repository: config.dirty_repository,
         })))
     }
@@ -213,53 +203,7 @@ impl Repository {
     /// already tried updating. It is therefore really important to call this
     /// method before doing any new update.
     pub fn start(&self) {
-        if let Some(ref rsync) = self.0.rsync {
-            rsync.0.lock().unwrap().clear_seen();
-        }
-    }
-
-    /// Asynchronously updates the content of the local repository.
-    ///
-    /// This goes over all the known repository publication points and runs
-    /// rsync on them to update.
-    ///
-    /// The method logs all error messages.
-    pub fn update_async(&self) -> impl Future<Item=(), Error=Error> {
-        let repo = self.clone();
-        let pool = CpuPool::new(self.0.rsync_threads);
-        fs::read_dir(&self.0.cache_dir).map_err(|err| {
-            error!(
-                "Failed to read repository directory: {}",
-                err
-            );
-            Error
-        }).into_future()
-        .and_then(|dir| {
-            future::join_all(dir.map(move |entry| {
-                let repo = repo.clone();
-                let pool = pool.clone();
-                entry.map_err(|err| {
-                    error!(
-                        "Failed to iterate over repository directory: {}",
-                        err
-                    );
-                    Error
-                }).into_future()
-                .and_then(move |entry| {
-                    pool.spawn(future::lazy(|| {
-                        repo.update_host(entry);
-                        Ok(())
-                    }))
-                })
-            })).map(|_| ())
-        })
-    }
-
-    /// Updates the content of the local repository.
-    ///
-    /// This is the synchronous version of `update_async`.
-    pub fn update(&self) -> Result<(), Error> {
-        self.update_async().wait()
+        self.0.rsync.start()
     }
 
     /// Process the local repository and produce a list of route origins.
@@ -280,11 +224,7 @@ impl Repository {
             let pool = pool.clone();
             pool.spawn(future::lazy(move || repo.process_tal(idx)))
         })).and_then(move |res| {
-            if !clean_repo.0.dirty_repository {
-                if let Some(rsync) = &clean_repo.0.rsync {
-                    clean_repo.cleanup(&unwrap!(rsync.0.lock()))?;
-                }
-            }
+            clean_repo.cleanup();
             Ok(res)
         })
     }
@@ -304,17 +244,9 @@ impl Repository {
         config.load_exceptions(self.0.extra_output)
     }
 
-    /// Returns the rsync metrics.
-    ///
-    /// This can only be done once per update as it replaces the current
-    /// metrics with an empty set.
-    pub fn take_rsync_metrics(&self) -> Vec<(uri::RsyncModule, RsyncMetrics)> {
-        if let Some(ref rsync) = self.0.rsync {
-            rsync.0.lock().unwrap().take_seen()
-        }
-        else {
-            Vec::new()
-        }
+    /// Update metrics for this validation run.
+    pub fn update_metrics(&self, metrics: &mut Metrics) {
+        self.0.rsync.update_metrics(metrics);
     }
 }
 
@@ -343,127 +275,7 @@ impl Repository {
         uri: &uri::Rsync,
         create: bool
     ) -> Option<Bytes> {
-        match File::open(self.uri_to_path(uri)) {
-            Ok(mut file) => {
-                if let Some(rsync) = self.0.rsync.as_ref() {
-                    unwrap!(rsync.0.lock()).mark_used(uri.module())
-                }
-                let mut data = Vec::new();
-                if let Err(err) = file.read_to_end(&mut data) {
-                    warn!(
-                        "Failed to read file '{}': {}",
-                        self.uri_to_path(uri).display(),
-                        err
-                    );
-                    return None
-                }
-                Some(data.into())
-            }
-            Err(ref err) if err.kind() == io::ErrorKind::NotFound => {
-                if create {
-                    self.rsync_module(uri.module());
-                    self.load_file(uri, false)
-                }
-                else {
-                    info!("{}: not found in local repository", uri);
-                    None
-                }
-            }
-            Err(err) => {
-                warn!(
-                    "Failed to open file '{}': {}",
-                    self.uri_to_path(uri).display(), err
-                );
-                None
-            }
-        }
-    }
-
-    /// Converts an rsync module URI into a path.
-    fn module_to_path(&self, module: &uri::RsyncModule) -> PathBuf {
-        let mut res = self.0.cache_dir.clone();
-        res.push(module.authority());
-        res.push(module.module());
-        res
-    }
-
-    /// Converts an rsync URI into a path.
-    fn uri_to_path(&self, uri: &uri::Rsync) -> PathBuf {
-        let mut res = self.module_to_path(uri.module());
-        res.push(uri.path());
-        res
-    }
-}
-
-
-/// # Updating
-///
-impl Repository {
-    /// Updates content of the host-specific directory of the local copy.
-    ///
-    /// If something goes wrong, logs a message.
-    fn update_host(self, entry: DirEntry) {
-        if !entry.file_type().map(|ft| ft.is_dir()).unwrap_or(false) {
-            return
-        }
-        match entry_to_uri_component(&entry) {
-            Some(host) => self.update_module(host, entry.path()),
-            None => {
-                warn!(
-                    "{}: illegal host directory. Skipping.",
-                    entry.path().display()
-                );
-            }
-        }
-    }
-
-    /// Updates content of the module-specific directory of the local copy.
-    ///
-    /// If something goes wrong, logs a message.
-    fn update_module(&self, host: Bytes, path: PathBuf) {
-        let dir = match fs::read_dir(&path) {
-            Ok(dir) => dir,
-            Err(err) => {
-                warn!(
-                    "Failed to read directory {}: {}.",
-                    path.display(), err
-                );
-                return
-            }
-        };
-        for entry in dir {
-            let entry = match entry {
-                Ok(entry) => entry,
-                Err(err) => {
-                    warn!(
-                        "Failed to iterate over directory {}: {}",
-                        path.display(), err
-                    );
-                    // XXX Or continue?
-                    return
-                }
-            };
-            if !entry.file_type().map(|t| t.is_dir()).unwrap_or(false) {
-                warn!(
-                    "{}: unexpected file. Skipping.",
-                    entry.path().display()
-                );
-                continue
-            }
-            match entry_to_uri_component(&entry) {
-                Some(module) => {
-                    self.rsync_module(
-                        &uri::RsyncModule::new(host.clone(), module)
-                    )
-                }
-                None => {
-                    warn!(
-                        "{}: illegal module directory. Skipping",
-                        entry.path().display()
-                    )
-                }
-            }
-        }
+        self.0.rsync.load_file(uri, create)
     }
 }
 
@@ -763,52 +575,22 @@ impl Repository {
 }
 
 
-/// # Cleaning up
+/// # Cleanup
 ///
 impl Repository {
-    fn cleanup(&self, rsync: &RsyncState) -> Result<(), Error> {
-        let dir = fs::read_dir(&self.0.cache_dir).map_err(|err| {
-            error!(
-                "Failed to read repository directory: {}",
-                err
-            );
-            Error
-        })?;
-        for entry in dir {
-            let entry = entry.map_err(|err| {
-                error!(
-                    "Failed to iterate over repository directory: {}",
-                    err
-                );
-                Error
-            })?;
-            self.cleanup_host(entry, rsync);
-        }
-        Ok(())
-    }
-
-    fn cleanup_host(&self, entry: DirEntry, rsync: &RsyncState) {
-        if !entry.file_type().map(|ft| ft.is_dir()).unwrap_or(false) {
+    fn cleanup(&self) {
+        if self.0.dirty_repository {
             return
         }
-        let path = entry.path();
-        let host = match entry_to_uri_component(&entry) {
-            Some(host) => host,
-            None => {
-                warn!(
-                    "{}: illegal host directory. Skipping.",
-                    path.display()
-                );
-                return
-            }
-        };
-        let dir = match fs::read_dir(&path) {
+        self.0.rsync.cleanup();
+        self.cleanup_base();
+    }
+
+    fn cleanup_base(&self) {
+        let dir = match fs::read_dir(&self.0.cache_dir) {
             Ok(dir) => dir,
             Err(err) => {
-                warn!(
-                    "Failed to read directory {}: {}. Skipping.",
-                    path.display(), err
-                );
+                warn!("Failed to read repository directory: {}", err);
                 return
             }
         };
@@ -817,399 +599,35 @@ impl Repository {
                 Ok(entry) => entry,
                 Err(err) => {
                     warn!(
-                        "Failed to iterate over directory {}: {}",
-                        path.display(), err
+                        "Failed to iterate over repository directory: {}", err
                     );
                     return
                 }
             };
-            if !entry.file_type().map(|t| t.is_dir()).unwrap_or(false) {
-                warn!(
-                    "{}: unexpected file. Skipping.",
-                    entry.path().display()
-                );
-                continue
+            match entry.file_name().to_str() {
+                Some("rsync") => continue,
+                _ => { }
             }
-            match entry_to_uri_component(&entry) {
-                Some(module) => {
-                    self.cleanup_module(
-                        uri::RsyncModule::new(host.clone(), module),
-                        entry.path(),
-                        rsync
-                    )
-                }
-                None => {
+            if entry.file_type().map(|t| t.is_dir()).unwrap_or(false) {
+                if let Err(err) = fs::remove_dir_all(entry.path()) {
                     warn!(
-                        "{}: illegal module directory. Skipping",
-                        entry.path().display()
-                    )
+                        "Failed to delete unused repository directory {}:{}",
+                        entry.path().display(),
+                        err
+                    );
                 }
             }
-        }
-        // Now we just try to delete the whole host directory which fails
-        // if there are still modules left.
-        let _ = fs::remove_dir(path);
-    }
-
-    fn cleanup_module(
-        &self,
-        module: uri::RsyncModule,
-        path: PathBuf,
-        rsync: &RsyncState
-    ) {
-        if !rsync.was_used(&module) {
-            debug!("Cleanup: trying to delete {}.", path.display());
-            if let Err(err) = fs::remove_dir_all(&path) {
-                error!(
-                    "Failed to delete rsync module directory {}: {}",
-                    path.display(),
-                    err
-                );
-            }
-        }
-        else {
-            debug!("Cleanup: keeping {}.", path.display());
-        }
-    }
-}
-
-
-/// # Rsyncing
-///
-impl Repository {
-    #[allow(clippy::mutex_atomic)] // XXX Double check maybe they are right?
-    fn rsync_module(&self, module: &uri::RsyncModule) {
-        if let Some((ref state, ref command)) = self.0.rsync {
-            if state.lock().unwrap().have_seen(module) {
-                return
-            }
-            let path = self.module_to_path(module);
-
-            let cvar = state.lock().unwrap().get_running(module);
-            match cvar {
-                Ok(cvar) => {
-                    let mut finished = cvar.0.lock().unwrap();
-                    while !*finished {
-                        finished = cvar.1.wait(finished).unwrap();
-                    }
-                }
-                Err(cvar) => {
-                    let mut finished = cvar.0.lock().unwrap();
-                    let metrics = command.update(module, path);
-                    {
-                        let mut state = state.lock().unwrap();
-                        state.remove_running(module);
-                        state.add_seen(module.clone(), metrics);
-                    }
-                    *finished = true;
-                    cvar.1.notify_all();
+            else {
+                if let Err(err) = fs::remove_file(entry.path()) {
+                    warn!(
+                        "Failed to delete unused repository entry {}:{}",
+                        entry.path().display(),
+                        err
+                    );
                 }
             }
         }
     }
-}
-
-
-//------------ RsyncState ----------------------------------------------------
-
-#[derive(Debug)]
-struct RsyncState {
-    /// Rsync processes currently running.
-    ///
-    /// The first element of each list item is the module for which the
-    /// process runs, the second is a conditional variable that is going
-    /// to be triggered when the process finishes.
-    #[allow(clippy::type_complexity)]
-    running: Vec<(uri::RsyncModule, Arc<(Mutex<bool>, Condvar)>)>,
-
-    /// The rsync modules we already tried in this iteration.
-    seen: HashMap<uri::RsyncModule, RsyncMetrics>,
-}
-
-impl RsyncState {
-    fn new() -> Self {
-        RsyncState {
-            running: Vec::new(),
-            seen: HashMap::new(),
-        }
-    }
-
-    #[allow(clippy::type_complexity, clippy::mutex_atomic)]
-    fn get_running(
-        &mut self,
-        module: &uri::RsyncModule
-    ) -> Result<Arc<(Mutex<bool>, Condvar)>, Arc<(Mutex<bool>, Condvar)>> {
-        for item in &self.running {
-            if item.0.eq(module) {
-                return Ok(item.1.clone())
-            }
-        }
-        let res = Arc::new((Mutex::new(false), Condvar::new()));
-        self.running.push((module.clone(), res.clone()));
-        Err(res)
-    }
-
-    fn remove_running(&mut self, module: &uri::RsyncModule) {
-        self.running.retain(|item| !item.0.eq(module))
-    }
-
-    fn add_seen(&mut self, module: uri::RsyncModule, metrics: RsyncMetrics) {
-        let _ = self.seen.insert(module, metrics);
-    }
-
-    fn mark_used(&mut self, module: &uri::RsyncModule) {
-        unwrap!(self.seen.get_mut(module)).used = true;
-    }
-
-    fn have_seen(&self, module: &uri::RsyncModule) -> bool {
-        self.seen.contains_key(module)
-    }
-
-    fn was_used(&self, module: &uri::RsyncModule) -> bool {
-        self.seen.get(module).map(|state| state.used).unwrap_or(false)
-    }
-
-    fn clear_seen(&mut self) {
-        self.seen.clear()
-    }
-
-    fn take_seen(&mut self) -> Vec<(uri::RsyncModule, RsyncMetrics)> {
-        mem::replace(&mut self.seen, HashMap::new()).into_iter().collect()
-    }
-}
-
-
-//------------ RsyncCommand --------------------------------------------------
-
-#[derive(Clone, Debug)]
-pub struct RsyncCommand {
-    command: String,
-    args: Vec<String>,
-    timeout: Duration,
-}
-
-impl RsyncCommand {
-    pub fn detect(config: &Config) -> Result<Self, Error> {
-        let command = config.rsync_command.clone();
-        let output = match process::Command::new(&command).arg("-h").output() {
-            Ok(output) => output,
-            Err(err) => {
-                error!(
-                    "Failed to run rsync: {}",
-                    err
-                );
-                return Err(Error)
-            }
-        };
-        if !output.status.success() {
-            error!(
-                "Running rsync failed with output: \n{}",
-                String::from_utf8_lossy(&output.stderr)
-            );
-            return Err(Error);
-        }
-        let args = match config.rsync_args {
-            Some(ref args) => args.clone(),
-            None => {
-                let has_contimeout =
-                   output.stdout.windows(12)
-                   .any(|window| window == b"--contimeout");
-                let timeout = format!(
-                    "--timeout={}",
-                    config.rsync_timeout.as_secs()
-                );
-                if has_contimeout {
-                    vec!["--contimeout=10".into(), timeout]
-                }
-                else {
-                    vec![timeout]
-                }
-            }
-        };
-        Ok(RsyncCommand {
-            command,
-            args,
-            timeout: config.rsync_timeout
-        })
-    }
-
-    pub fn update<P: AsRef<Path>>(
-        &self,
-        source: &uri::RsyncModule,
-        destination: P
-    ) -> RsyncMetrics {
-        let start = SystemTime::now();
-        let status = {
-            match self.command(source, destination) {
-                Ok(mut command) => match command.output() {
-                    Ok(output) => Ok(Self::log_output(source, output)),
-                    Err(err) => Err(err)
-                }
-                Err(err) => Err(err)
-            }
-        };
-        RsyncMetrics {
-            status,
-            duration: SystemTime::now().duration_since(start),
-            used: false,
-        }
-    }
-
-    fn command<P: AsRef<Path>>(
-        &self,
-        source: &uri::RsyncModule,
-        destination: P
-    ) -> Result<process::Command, io::Error> {
-        info!("rsyncing from {}.", source);
-        let destination = destination.as_ref();
-        create_dir_all(destination)?;
-        let destination = match Self::format_destination(destination) {
-            Ok(some) => some,
-            Err(_) => {
-                error!(
-                    "rsync: illegal destination path {}.",
-                    destination.display()
-                );
-                return Err(io::Error::new(
-                    io::ErrorKind::Other,
-                    "illegal destination path"
-                ));
-            }
-        };
-        let mut cmd = process::Command::new(&self.command);
-        for item in &self.args {
-            cmd.arg(item);
-        }
-        cmd.arg("-rltz")
-           .arg("--delete")
-           .arg(source.to_string())
-           .arg(destination);
-        debug!(
-            "rsync://{}/{}: Running command {:?}",
-            source.authority(), source.module(), cmd
-        );
-        Ok(cmd)
-    }
-
-    #[cfg(not(windows))]
-    fn format_destination(path: &Path) -> Result<String, Error> {
-        let mut destination = format!("{}", path.display());
-        if !destination.ends_with('/') {
-            destination.push('/')
-        }
-        Ok(destination)
-    }
-
-    #[cfg(windows)]
-    fn format_destination(path: &Path) -> Result<String, Error> {
-        // On Windows we are using Cygwin rsync which requires Unix-style
-        // paths. In particular, the drive parameter needs to be turned
-        // from e.g. `C:` into `/cygdrive/c` and all backslashes should
-        // become slashes.
-        use std::path::{Component, Prefix};
-
-        let mut destination = String::new();
-        for component in path.components() {
-            match component {
-                Component::Prefix(prefix) => {
-                    // We only accept UNC and Disk prefixes. Everything else
-                    // causes an error.
-                    match prefix.kind() {
-                        Prefix::UNC(server, share) => {
-                            let (server, share) = match (server.to_str(),
-                                                         share.to_str()) {
-                                (Some(srv), Some(shr)) => (srv, shr),
-                                _ => return Err(Error)
-                            };
-                            destination.push_str(server);
-                            destination.push('/');
-                            destination.push_str(share);
-                        }
-                        Prefix::Disk(disk) => {
-                            let disk = if disk.is_ascii() {
-                                (disk as char).to_ascii_lowercase()
-                            }
-                            else {
-                                return Err(Error)
-                            };
-                            destination.push_str("/cygdrive/");
-                            destination.push(disk);
-                        }
-                        _ => return Err(Error)
-                    }
-                }
-                Component::CurDir | Component::RootDir => {
-                    continue
-                }
-                Component::ParentDir => {
-                    destination.push_str("..");
-                }
-                Component::Normal(s) => {
-                    match s.to_str() {
-                        Some(s) => destination.push_str(s),
-                        None => return Err(Error)
-                    }
-                }
-            }
-            destination.push_str("/");
-        }
-        Ok(destination)
-    }
-
-    fn log_output(source: &uri::RsyncModule, output: Output) -> ExitStatus {
-        if !output.status.success() {
-            warn!(
-                "rsync://{}/{}: failed with status {}",
-                source.authority(), source.module(), output.status
-            );
-        }
-        else {
-            debug!(
-                "rsync://{}/{}: successfully completed.",
-                source.authority(), source.module(),
-            );
-        }
-        if !output.stderr.is_empty() {
-            String::from_utf8_lossy(&output.stderr).lines().for_each(|l| {
-                warn!(
-                    "rsync://{}/{}: {}", source.authority(), source.module(), l
-                );
-            })
-        }
-        if !output.stdout.is_empty() {
-            String::from_utf8_lossy(&output.stdout).lines().for_each(|l| {
-                info!(
-                    "rsync://{}/{}: {}", source.authority(), source.module(), l
-                )
-            })
-        }
-        output.status
-    }
-}
-
-
-//------------ RsyncMetrics --------------------------------------------------
-
-#[derive(Debug)]
-pub struct RsyncMetrics {
-    pub status: Result<ExitStatus, io::Error>,
-    pub duration: Result<Duration, SystemTimeError>,
-    pub used: bool,
-}
-
-
-//------------ Helper Functions ----------------------------------------------
-
-fn entry_to_uri_component(entry: &DirEntry) -> Option<Bytes> {
-    let name = entry.file_name();
-    name.to_str().and_then(|name| {
-        if uri::is_uri_ascii(name) {
-            Some(Bytes::from(name.as_bytes()))
-        }
-        else {
-            None
-        }
-    })
 }
